requirements.txt
numpy
pillow
matplotlib
argparse